GH-988: move formfield and add uuid
