+++ ui/package.json
@@ -41,6 +41,7 @@
     "redux-persist": "^6.0.0",
     "sass": "^1.53.0",
     "slick-carousel": "^1.8.1",
+    "uuid": "^9.0.1",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
